+++ strutil/strutil_test.go
@@ -124,11 +124,15 @@ func (s *StrUtilSuite) TestFields(c *C) {
 }
 
 func (s *StrUtilSuite) TestReadField(c *C) {
-	c.Assert(ReadField("abc 1234 DEF", -1), Equals, "")
-	c.Assert(ReadField("abc 1234 DEF", 0), Equals, "abc")
-	c.Assert(ReadField("abc 1234 DEF", 1), Equals, "1234")
-	c.Assert(ReadField("abc 1234 DEF", 2), Equals, "DEF")
-	c.Assert(ReadField("abc 1234 DEF", 3), Equals, "")
+	c.Assert(ReadField("abc 1234 DEF", -1, true), Equals, "")
+	c.Assert(ReadField("abc 1234 DEF", 0, true), Equals, "abc")
+	c.Assert(ReadField("abc 1234 DEF", 1, true), Equals, "1234")
+	c.Assert(ReadField("abc 1234 DEF", 2, true), Equals, "DEF")
+	c.Assert(ReadField("abc 1234 DEF", 3, true), Equals, "")
+
+	c.Assert(ReadField("abc|||||1234||DEF", 1, true, "|"), Equals, "1234")
+	c.Assert(ReadField("abc+1234|DEF", 1, true, "|+"), Equals, "1234")
+	c.Assert(ReadField("abc::1234:::DEF:", 5, false, ":"), Equals, "DEF")
 }
 
 func (s *StrUtilSuite) BenchmarkFields(c *C) {
@@ -139,7 +143,7 @@ func (s *StrUtilSuite) BenchmarkFields(c *C) {
 
 func (s *StrUtilSuite) BenchmarkReadField(c *C) {
 	for i := 0; i < c.N; i++ {
-		ReadField("abc 1234 DEF", 2)
+		ReadField("abc 1234 DEF", 2, false)
 	}
 }